requirements.txt
torch
torchvision
torchaudio
transformers
datasets
tokenizers
sentencepiece
numpy
pandas
matplotlib
scikit-learn
tqdm
pyyaml
rich
einops
pytest
wandb
